+++ CHANGES.md
@@ -1,6 +1,8 @@
 #### [unreleased]
 * use `afragen/wp-readme-parser` drop-in replacement for `afragen/wordpress-plugin-readme-parser`
 * update requirements to PHP 8.1 for new parser due to testing
+* add `maybe_extend_repo_cache()` to update the timeout if the remote and cached version numbers are same, should avoid API calls for current data
+* update `(get|set)_repo_cache()`
 
 #### 12.24.2 / 2026-03-25
 * update freemius/wordpress-sdk